--- Manito/Manito/Global/UIComponent/ToastPopupView.swift
@@ -1,14 +1,0 @@
-//
-//  ToastPopupView.swift
-//  Manito
-//
-//  Created by 이성호 on 2022/11/14.
-//
-
-import UIKit
-
-import SnapKit
-
-class ToastPopupView: UIView {
-    
-}
